RabbitMQ UI not fixed
I have been attempting to fix the RabbitMQ UI issue for a few days and have no been able to fix the issue.

A bunch of code changed and added a dockerfile to help the container startup quicker, removed celery_worker.py and used entrypoint.sh instead.

For now I will move on from it and visit later in order to continue.